Refactor client handlers
