+++ instructor/client_bedrock.py
@@ -1,6 +1,10 @@
 from __future__ import annotations  # type: ignore
 
-from typing import Any, Literal, overload
+from typing import Any, Literal, overload, TypeVar
+import json
+from textwrap import dedent
+
+T = TypeVar("T")
 
 import boto3
 from botocore.client import BaseClient
@@ -25,15 +29,6 @@ def from_bedrock(
     _async: Literal[True] = True,
     **kwargs: Any,
 ) -> AsyncInstructor: ...
-
-
-def handle_bedrock_json(
-    response_model: Any,
-    new_kwargs: Any,
-) -> tuple[Any, Any]:
-    print(f"handle_bedrock_json: response_model {response_model}")
-    print(f"handle_bedrock_json: new_kwargs {new_kwargs}")
-    return response_model, new_kwargs
 
 
 def from_bedrock(
@@ -85,3 +80,125 @@ def from_bedrock(
             mode=mode,
             **kwargs,
         )
+
+
+def _prepare_bedrock_converse_kwargs_internal(
+    call_kwargs: dict[str, Any],
+) -> dict[str, Any]:
+    """Minimal processing to support `converse` parameters for the Bedrock client
+
+    See: https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/bedrock-runtime/client/converse.html
+    """
+    if "model" in call_kwargs and "modelId" not in call_kwargs:
+        call_kwargs["modelId"] = call_kwargs.pop("model")
+
+    inference_config_params = {}
+
+    if "temperature" in call_kwargs:
+        inference_config_params["temperature"] = call_kwargs.pop("temperature")
+
+    if "max_tokens" in call_kwargs:
+        inference_config_params["maxTokens"] = call_kwargs.pop("max_tokens")
+    elif "maxTokens" in call_kwargs:
+        inference_config_params["maxTokens"] = call_kwargs.pop("maxTokens")
+
+    if "top_p" in call_kwargs:
+        inference_config_params["topP"] = call_kwargs.pop("top_p")
+    elif "topP" in call_kwargs:
+        inference_config_params["topP"] = call_kwargs.pop("topP")
+
+    if "stop" in call_kwargs:
+        stop_val = call_kwargs.pop("stop")
+        if isinstance(stop_val, str):
+            inference_config_params["stopSequences"] = [stop_val]
+        elif isinstance(stop_val, list):
+            inference_config_params["stopSequences"] = stop_val
+    elif "stop_sequences" in call_kwargs:
+        inference_config_params["stopSequences"] = call_kwargs.pop("stop_sequences")
+    elif "stopSequences" in call_kwargs:
+        inference_config_params["stopSequences"] = call_kwargs.pop("stopSequences")
+
+    if inference_config_params:
+        if "inferenceConfig" in call_kwargs:
+            existing_inference_config = call_kwargs["inferenceConfig"]
+            for key, value in inference_config_params.items():
+                if key not in existing_inference_config:
+                    existing_inference_config[key] = value
+        else:
+            call_kwargs["inferenceConfig"] = inference_config_params
+
+    if "messages" in call_kwargs and isinstance(call_kwargs["messages"], list):
+        original_input_messages = call_kwargs.pop("messages")
+
+        bedrock_system_list: list[dict[str, Any]] = []
+        bedrock_user_assistant_messages_list: list[dict[str, Any]] = []
+
+        for msg_dict in original_input_messages:
+            if not isinstance(msg_dict, dict):
+                bedrock_user_assistant_messages_list.append(msg_dict)
+                continue
+
+            current_message_for_api = msg_dict.copy()
+            role = current_message_for_api.get("role")
+            content = current_message_for_api.get("content")
+
+            if role == "system":
+                if isinstance(content, str):
+                    bedrock_system_list.append({"text": content})
+                else:
+                    raise ValueError(
+                        "System message content must be a string for Bedrock processing by this handler. "
+                        f"Found type: {type(content)}."
+                    )
+            else:
+                if "content" in current_message_for_api:
+                    if isinstance(content, str):
+                        current_message_for_api["content"] = [{"text": content}]
+                    else:
+                        raise NotImplementedError(
+                            "Non-text prompts are not currently supported in the Bedrock provider."
+                        )
+                bedrock_user_assistant_messages_list.append(current_message_for_api)
+
+        if bedrock_system_list:
+            call_kwargs["system"] = bedrock_system_list
+
+        call_kwargs["messages"] = bedrock_user_assistant_messages_list
+    return call_kwargs
+
+
+def handle_bedrock_json(
+    response_model: type[T], new_kwargs: dict[str, Any]
+) -> tuple[type[T], dict[str, Any]]:
+    new_kwargs = _prepare_bedrock_converse_kwargs_internal(new_kwargs)
+    json_message = dedent(
+        f"""
+        As a genius expert, your task is to understand the content and provide
+        the parsed objects in json that match the following json_schema:\n
+        {json.dumps(response_model.model_json_schema(), indent=2, ensure_ascii=False)}
+
+        Make sure to return an instance of the JSON, not the schema itself
+        and don't include any other text in the response apart from the json
+        """
+    )
+    system_message = new_kwargs.pop("system", None)
+    if not system_message:
+        new_kwargs["system"] = [{"text": json_message}]
+    else:
+        if not isinstance(system_message, list):
+            raise ValueError(
+                """system must be a list of SystemMessage, refer to:
+                https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/bedrock-runtime/client/converse.html
+                """
+            )
+        system_message.append({"text": json_message})
+        new_kwargs["system"] = system_message
+
+    return response_model, new_kwargs
+
+
+def handle_bedrock_tools(
+    response_model: type[T], new_kwargs: dict[str, Any]
+) -> tuple[type[T], dict[str, Any]]:
+    new_kwargs = _prepare_bedrock_converse_kwargs_internal(new_kwargs)
+    return response_model, new_kwargs
