+++ instructor/client_cerebras.py
@@ -1,6 +1,8 @@
 from __future__ import annotations  # type: ignore
 
-from typing import Any, overload
+from typing import Any, overload, TypeVar
+
+T = TypeVar("T")
 
 import instructor
 from instructor.client import AsyncInstructor, Instructor
@@ -70,3 +72,41 @@ def from_cerebras(
         mode=mode,
         **kwargs,
     )
+
+
+def handle_cerebras_tools(
+    response_model: type[T], new_kwargs: dict[str, Any]
+) -> tuple[type[T], dict[str, Any]]:
+    if new_kwargs.get("stream", False):
+        raise ValueError("Stream is not supported for Cerebras Tool Calling")
+    new_kwargs["tools"] = [
+        {
+            "type": "function",
+            "function": response_model.openai_schema,
+        }
+    ]
+    new_kwargs["tool_choice"] = {
+        "type": "function",
+        "function": {"name": response_model.openai_schema["name"]},
+    }
+    return response_model, new_kwargs
+
+
+def handle_cerebras_json(
+    response_model: type[T], new_kwargs: dict[str, Any]
+) -> tuple[type[T], dict[str, Any]]:
+    instruction = f"""
+You are a helpful assistant that excels at following instructions.Your task is to understand the content and provide the parsed objects in json that match the following json_schema:\n
+Here is the relevant JSON schema to adhere to
+
+<schema>
+{response_model.model_json_schema()}
+</schema>
+
+Your response should consist only of a valid JSON object that `{response_model.__name__}.model_validate_json()` can successfully parse.
+"""
+
+    new_kwargs["messages"] = [{"role": "system", "content": instruction}] + new_kwargs[
+        "messages"
+    ]
+    return response_model, new_kwargs
