+++ instructor/client_writer.py
@@ -4,7 +4,11 @@ from __future__ import annotations
 
 import instructor
 from writerai import AsyncWriter, Writer
-from typing import overload, Any
+from typing import overload, Any, TypeVar
+
+from .openai_utils import handle_openai_json_schema, handle_openai_tools
+
+T = TypeVar("T")
 
 
 @overload
@@ -61,3 +65,19 @@ def from_writer(
         mode=mode,
         **kwargs,
     )
+
+
+def handle_writer_tools(
+    response_model: type[T], new_kwargs: dict[str, Any]
+) -> tuple[type[T], dict[str, Any]]:
+    return handle_openai_tools(
+        response_model,
+        new_kwargs,
+        tool_choice="auto",
+    )
+
+
+def handle_writer_json(
+    response_model: type[T], new_kwargs: dict[str, Any]
+) -> tuple[type[T], dict[str, Any]]:
+    return handle_openai_json_schema(response_model, new_kwargs)
